reorganized code into opmodes, samples, and test directories
