+++ 2021/samples/SkyStoneAuto_OneSkystone_sample.java
@@ -1,4 +1,4 @@
-package test;
+package samples;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.hardware.DcMotor;
